standardized component/screen/page naming conventions
